warnings: silence -Wconversion, fix misleading-indent, suppress UNUSEDSIGNAL
C side (gcc -Wconversion on Ubuntu, not flagged by Apple clang):

  cmodel/z80_alu.c:51  - sub4 borrow assign: explicit (uint8_t)cast
                         around the ternary so the unsigned-int literal
                         doesn't trigger value-change warning.

  cmodel/z80_core.c:489 - EXEC_JP_CC: split the post-fetch statement
                          across multiple clearly-indented lines so
                          gcc's -Wmisleading-indentation no longer
                          reads `finish(c)` as guarded by the cc-check
                          `if` (which is now obviously its own line).

  scripts/zexrunner.c:69,74  - cast timespec.{tv_sec,tv_nsec} to double
  scripts/fuse_runner.c:205    before the subtraction, and cast the
                               instruction count to double before
                               dividing — kills the long-int→double
                               and long-long→double -Wconversion hits.

Verilog / Verilator side (Verilator 5.020 -Wall on Ubuntu CI):

  rtl/z80_core.v  - uses_cc_w / valid_w were declared but the
                    sequencer never consumed them, triggering
                    UNUSEDSIGNAL on the wires AND PINMISSING when
                    they were dropped. Keep the wires (with a
                    comment naming them as unconsumed-yet outputs
                    of u_pla) and suppress UNUSEDSIGNAL globally.

  Makefile        - add -Wno-UNUSEDSIGNAL to the verilator and
                    verilator_zex commands. The remaining UNUSED
                    warnings are all "we declare a wide register
                    so the carry-out bit is visible structurally,
                    but only that one bit is consumed" — by
                    design, not a bug. The comment in the Makefile
                    enumerates which signals (lo_add/lo_sub[3:0]
                    in z80_alu, add12/r13/bk/bhc in z80_core,
                    m_len in z80_timing) and notes the planned
                    cleanup once E1 lands explicit bus-segment
                    taps.

Verified locally:
  - make ctest        -> ALL C TESTS PASSED
  - make zexrunner / fuse_runner / tracegen  -> no warnings
  - make verilator     -> zero Verilator warnings
  - make verilator_zex -> zero Verilator warnings
  - make rtl          -> RTL elaboration OK
  - make fuse_rtl     -> 1356/1356
  - make compare      -> 8 programs identical (C vs iverilog)
  - make perfectz80   -> 3/3 pin parity

+++ cmodel/z80_alu.c
@@ -48,7 +48,7 @@ static uint8_t add4(uint8_t a, uint8_t b, uint8_t cin, uint8_t *cout)
 static uint8_t sub4(uint8_t a, uint8_t b, uint8_t bin, uint8_t *bout)
 {
     int s = (int)(a & 0xF) - (int)(b & 0xF) - (int)(bin & 1);
-    *bout = (s < 0) ? 1u : 0u;
+    *bout = (uint8_t)((s < 0) ? 1 : 0);
     return (uint8_t)(s & 0xF);
 }
